+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 215:
 * Remove bind_back_handler
 * bind_front_handler works with member functions
 * Examples use bind_front_handler
+* Add experimental test/handler.hpp
 
 --------------------------------------------------------------------------------